Refactoring telemetry initialization.
