+++ pkg/definitions/src/index.ts
@@ -6,4 +6,4 @@ export * from './deferredPromise';
 export * from './serviceRegistration';
 export * from './serviceBindings';
 export * from './useRemoteServices';
-export * from './logging';
+export * from './telemetry';
