--- pkg/definitions/src/logging.ts
@@ -1,20 +1,0 @@
-import {
-  ConsoleSpanExporter,
-  SimpleSpanProcessor,
-} from '@opentelemetry/sdk-trace-base';
-import { WebTracerProvider } from '@opentelemetry/sdk-trace-web';
-import { ZipkinExporter } from '@opentelemetry/exporter-zipkin';
-
-const provider = new WebTracerProvider();
-provider.addSpanProcessor(new SimpleSpanProcessor(new ConsoleSpanExporter()));
-provider.addSpanProcessor(new SimpleSpanProcessor(new ZipkinExporter({})));
-
-provider.register();
-
-export function getTracer() {
-  return provider.getTracer('shared-worker-service-invocation');
-}
-
-export function fromError(err: unknown) {
-  return err instanceof Error ? err : new Error(String(err));
-}
