+++ pkg/webapp/src/app/workers/sharedWorkerOne.ts
@@ -1,26 +1,39 @@
 import {
   exposeServiceOnPort,
+  fromError,
   mapRemoteServiceOnPort,
 } from '@example/definitions';
 import { createServiceA } from '@example/service-a';
 import * as Comlink from 'comlink';
+import { initializeTelemetry, getTracer } from '@example/definitions';
 
-const ctx = globalThis as unknown as SharedWorkerGlobalScope;
+initializeTelemetry({
+  serviceName: 'shared-worker-one',
+});
 
-const serviceA = createServiceA();
+const span = getTracer().startSpan('Initialize');
+try {
+  const ctx = globalThis as unknown as SharedWorkerGlobalScope;
 
-ctx.onconnect = (evt) => {
-  const [port] = evt.ports;
-  Comlink.expose(
-    {
-      exposeServiceOnPort: exposeServiceOnPort({
-        'a-service': serviceA,
-      }),
-      mapRemoteServiceOnPort: mapRemoteServiceOnPort(),
-      workerInfo: () => ({
-        title: 'Worker One',
-      }),
-    },
-    port
-  );
-};
+  const serviceA = createServiceA();
+
+  ctx.onconnect = (evt) => {
+    const [port] = evt.ports;
+    Comlink.expose(
+      {
+        exposeServiceOnPort: exposeServiceOnPort({
+          'a-service': serviceA,
+        }),
+        mapRemoteServiceOnPort: mapRemoteServiceOnPort(),
+        workerInfo: () => ({
+          title: 'Worker One',
+        }),
+      },
+      port
+    );
+  };
+} catch (error) {
+  span.recordException(fromError(error));
+}
+
+span.end();
