+++ pkg/webapp/src/main.tsx
@@ -2,13 +2,17 @@ import { StrictMode } from 'react';
 import * as ReactDOM from 'react-dom/client';
 import { connectToSharedWorkers } from './app/connectSharedWorkers';
 import App from './app/app';
-import { fromError, getTracer } from '@example/definitions';
+import { initializeTelemetry, fromError, getTracer } from '@example/definitions';
+
+initializeTelemetry({
+  serviceName: 'web-app',
+});
 
 initializeApp();
 
 async function initializeApp() {
   const tracer = getTracer();
-  const span = tracer.startSpan('Initialize App');
+  const span = tracer.startSpan('Initialize');
   try {
     await connectToSharedWorkers();
